backend/requirements.txt
fastapi
uvicorn[standard]
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-chroma>=0.2.0
langgraph>=0.2.0
chromadb>=0.5.0
pymupdf>=1.24.0
python-dotenv
pydantic>=2.0.0
